refactor(hook): add useGetLocation hook
