+++ apps/bus/src/components/Map.tsx
@@ -1,107 +1,18 @@
-import React, { useEffect, useRef } from 'react';
+import React, { useEffect } from 'react';
 
 import { Box, IconButton } from '@chakra-ui/react';
-import type mapboxgl from 'mapbox-gl';
 import Head from 'next/head';
 import Image from 'next/image';
 import { IoLocate } from 'react-icons/io5';
 
 import { useMap } from './MapContextProvider';
 
-import useAppToast from '@/hooks/use-app-toast';
+import useGetLocation from '@/hooks/use-get-location';
 import background from '@/map.jpg';
 
-const DEFAULT_ZOOM = 15;
-
 const Map = () => {
-  const toast = useAppToast();
   const { mapContextRef, divRef, isLoaded } = useMap();
-  const currentPositionRef = useRef<{ lat: number; lng: number } | null>(null);
-
-  const onLocate = async () => {
-    if (!window.navigator) {
-      toast({ description: '偵測定位系統失敗', status: 'error' });
-      return;
-    }
-
-    try {
-      const geoLocation = await new Promise<GeolocationPosition>(
-        (resolve, reject) =>
-          window.navigator.geolocation.getCurrentPosition(resolve, reject),
-      );
-
-      toast({ description: '成功獲取定位' });
-
-      console.warn(geoLocation);
-
-      const newPosition = {
-        center: [
-          geoLocation.coords.longitude,
-          geoLocation.coords.latitude,
-        ] as mapboxgl.LngLatLike,
-        zoom: DEFAULT_ZOOM,
-      };
-
-      const flyToCurrent = () => {
-        mapContextRef.current.map.flyTo({
-          center: newPosition.center,
-          zoom: DEFAULT_ZOOM,
-        });
-      };
-
-      currentPositionRef.current = {
-        lat: geoLocation.coords.latitude,
-        lng: geoLocation.coords.longitude,
-      };
-
-      const { createJSXMarker: attachJSXMarker } = await import(
-        '@/services/mapbox'
-      );
-
-      if (mapContextRef.current.map && mapContextRef.current.positionMarker) {
-        flyToCurrent();
-        mapContextRef.current.positionMarker.remove();
-      }
-
-      mapContextRef.current.positionMarker = attachJSXMarker(
-        <Box
-          id="current"
-          h="40px"
-          w="40px"
-          bgImage="url(/current.png)"
-          pointer="cursor"
-          onClick={flyToCurrent}
-          zIndex="modal"
-        />,
-        newPosition.center,
-      );
-      mapContextRef.current.positionMarker.addTo(mapContextRef.current.map);
-      flyToCurrent();
-    } catch (error) {
-      const { code } = error as GeolocationPositionError;
-      console.error(error);
-
-      let description: string;
-      switch (code) {
-        case GeolocationPositionError.PERMISSION_DENIED:
-          description = '授權定位失敗';
-          break;
-
-        case GeolocationPositionError.POSITION_UNAVAILABLE:
-          description = '該地區無定位服務';
-          break;
-
-        case GeolocationPositionError.TIMEOUT:
-          description = '獲取地址過久，連線中斷';
-          break;
-
-        default:
-          description = '未知錯誤';
-          break;
-      }
-      toast({ description, status: 'error' });
-    }
-  };
+  const { onLocate, currentPositionRef } = useGetLocation();
 
   useEffect(() => {
     if (!isLoaded || !mapContextRef.current.map) {
@@ -130,7 +41,7 @@ const Map = () => {
       map.off('render', handleMapLoad);
       map.off('move', handleMapMove);
     };
-  }, [isLoaded, mapContextRef]);
+  }, [currentPositionRef, isLoaded, mapContextRef]);
 
   return (
     <>
